Tweak derive_error
In the docs, data_store_error inherits derive_error instead of using
error to avoid manually defining the error alias template.  In the
derive_error implementation, it makes more sense to use protected for
the error type as it is not meant to be used only by derived classes.
Also, use class, not struct, as this is for inheritance.  On the
userside context, data_store_error uses struct rather than class as
it's just a container for error types.